+++ x/stake/store_test.go
@@ -12,7 +12,6 @@ import (
 	crypto "github.com/tendermint/go-crypto"
 	"github.com/tendermint/go-wire"
 	dbm "github.com/tendermint/tmlibs/db"
-	"github.com/tendermint/tmlibs/rational"
 )
 
 func newPubKey(pk string) (res crypto.PubKey, err error) {
@@ -57,9 +56,9 @@ func TestState(t *testing.T) {
 	candidate := &Candidate{
 		Owner:       validator,
 		PubKey:      pk,
-		Assets:      rational.New(9),
-		Liabilities: rational.New(9),
-		VotingPower: rational.Zero,
+		Assets:      9, //rational.New(9),
+		Liabilities: 9, // rational.New(9),
+		VotingPower: 0, //rational.Zero,
 	}
 
 	candidatesEqual := func(c1, c2 *Candidate) bool {
@@ -84,7 +83,7 @@ func TestState(t *testing.T) {
 	assert.True(candidatesEqual(candidate, resCand))
 
 	// modify a records, save, and retrieve
-	candidate.Liabilities = rational.New(99)
+	candidate.Liabilities = 99 //rational.New(99)
 	saveCandidate(store, candidate)
 	resCand = loadCandidate(store, pk)
 	assert.True(candidatesEqual(candidate, resCand))
@@ -99,7 +98,7 @@ func TestState(t *testing.T) {
 
 	bond := &DelegatorBond{
 		PubKey: pk,
-		Shares: rational.New(9),
+		Shares: 9, // rational.New(9),
 	}
 
 	bondsEqual := func(b1, b2 *DelegatorBond) bool {
@@ -117,7 +116,7 @@ func TestState(t *testing.T) {
 	assert.True(bondsEqual(bond, resBond))
 
 	//modify a records, save, and retrieve
-	bond.Shares = rational.New(99)
+	bond.Shares = 99 //rational.New(99)
 	saveDelegatorBond(store, delegator, bond)
 	resBond = loadDelegatorBond(store, delegator, pk)
 	assert.True(bondsEqual(bond, resBond))
